# Conflicts:
#	src/Nethermind/Nethermind.Api/IApiWithBlockchain.cs
#	src/Nethermind/Nethermind.Blockchain/Producers/DevBlockProducer.cs
#	src/Nethermind/Nethermind.HealthChecks/HealthChecksPlugin.cs
#	src/Nethermind/Nethermind.HealthChecks/Nethermind.HealthChecks.csproj
#	src/Nethermind/Nethermind.HealthChecks/NodeHealthChecks.cs
#	src/Nethermind/Nethermind.JsonRpc.Test/Modules/TestRpcBlockchain.cs
#	src/Nethermind/Nethermind.Runner/Ethereum/Steps/InitializeBlockchain.cs
#	src/Nethermind/Nethermind.Runner/Ethereum/Steps/InitializeBlockchainAuRa.cs
#	src/Nethermind/Nethermind.sln

+++ src/Nethermind/Nethermind.Baseline/Config/BaselineConfig.cs
@@ -23,13 +23,13 @@ namespace Nethermind.Baseline.Config
         public bool Enabled { get; set; }
 
         public bool BaselineTreeDbCacheIndexAndFilterBlocks { get; set; } = false;
-        public ulong BaselineTreeDbBlockCacheSize { get; set; } = 1.KiB();
-        public ulong BaselineTreeDbWriteBufferSize { get; set; } = 1.KiB();
+        public ulong BaselineTreeDbBlockCacheSize { get; set; } = (ulong)1.KiB();
+        public ulong BaselineTreeDbWriteBufferSize { get; set; } = (ulong)1.KiB();
         public uint BaselineTreeDbWriteBufferNumber { get; set; } = 4;
 
         public bool BaselineTreeMetadataDbCacheIndexAndFilterBlocks { get; set; } = false;
-        public ulong BaselineTreeMetadataDbBlockCacheSize { get; set; } = 1.KiB();
-        public ulong BaselineTreeMetadataDbWriteBufferSize { get; set; } = 1.KiB();
+        public ulong BaselineTreeMetadataDbBlockCacheSize { get; set; } = (ulong)1.KiB();
+        public ulong BaselineTreeMetadataDbWriteBufferSize { get; set; } = (ulong)1.KiB();
         public uint BaselineTreeMetadataDbWriteBufferNumber { get; set; } = 4;
     }
 }
